refactor: 刪除不再使用的元件 CategoryCard、FollowButton 和 RecipeCard，並更新相關引入路徑以反映新結構，提升專案的可維護性和清晰度。
