+++ src/components/pages/AuthorProfile/AuthorRecipes.tsx
@@ -1,6 +1,6 @@
 import { ChevronDown } from 'lucide-react';
 import { useState, useEffect } from 'react';
-import { RecipeCard } from '@/components/common/RecipeCard';
+import { RecipeCard } from '@/components/features/RecipeCard';
 import { fetchUserRecipes } from '@/services/users';
 
 const apiBaseUrl = process.env.NEXT_PUBLIC_API_BASE_URL_DEV;
