+++ src/pages/index.tsx
@@ -4,8 +4,8 @@ import { Plus, ChevronDown, X } from 'lucide-react';
 // import { Header } from '@/components/layout/Header';
 import { Button } from '@/components/ui/button';
 import { Tabs, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { CategoryCard } from '@/components/common/CategoryCard';
-import { RecipeCard } from '@/components/common/RecipeCard';
+import { CategoryCard } from '@/components/features/CategoryCard';
+import { RecipeCard } from '@/components/features/RecipeCard';
 import { Carousel } from '@/components/ui/carousel';
 import { useRouter } from 'next/router';
 import { GetStaticProps } from 'next';
